requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
